+++ @navikt/core/react/src/modal/index.ts
@@ -1,4 +1,5 @@
-export { default as Modal, type ModalProps } from "./Modal";
-export { type ModalHeaderProps } from "./ModalHeader";
+export { default as Modal } from "./Modal";
 export { type ModalBodyProps } from "./ModalBody";
 export { type ModalFooterProps } from "./ModalFooter";
+export { type ModalHeaderProps } from "./ModalHeader";
+export { type ModalProps } from "./types";
